pyproject.toml
[project]
name = "tp-jv"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2>=2.9.11",
]
